pyproject.toml
[project]
name = "lab-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "opencv-contrib-python>=4.13.0.92",
    "pillow-avif-plugin>=1.5.5",
    "pyqt6>=6.10.2",
    "rawpy>=0.26.1",
    "scikit-image>=0.26.0",
]
